play with types, interfaces versus implmentation
