+++ clients/webclients/utils/bang/references/render.ts
@@ -1,3 +1,8 @@
+// brian taylor vann
+
+// N Node
+// A Attributables
+
 interface RenderResults<A> {
   templateArray: TemplateStringsArray;
   injections: A[];
